Remove Acme Next workspace and just make it the default workspace
